requirements.txt
pytest
coverage
black
isort
nox
flask
flask-restful
gunicorn
django
django-extensions
psycopg2-binary